+++ Tests/PDKitTests/Pills/PillAttributeTests.swift
@@ -1,0 +1,9 @@
+//
+//  PillAttributeTests.swift
+//  PatchDay
+//
+//  Created by Juliya Smith on 2/10/21.
+//  Copyright © 2021 Juliya Smith. All rights reserved.
+//
+
+import Foundation